Marvel App
It's a Marvel app that I coded with React-Native.
Marvel API is used in the application.
The user can register and log in with his e-mail address.
In the application, the user can view the Marvel characters and information such as the comics with those characters. Likewise, it can display a comic book and the characters in that comic.
There is also a comic book detail or character detail screen in the application. You can see the characters, stories, etc. in the comic book detail screen.
There is information such as comics and stories in the character detail.